requirements.txt
FastAPI
uvicorn
pydantic
os
json
config
google.generativeai
